+++ server/src/test/java/org/opensearch/action/admin/cluster/stats/ClusterStatsNodesTests.java
@@ -316,7 +316,6 @@ public class ClusterStatsNodesTests extends OpenSearchTestCase {
             defaultClusterStatsIndices.getSegments().getVersionMapMemoryInBytes(),
             optimzedClusterStatsIndices.getSegments().getVersionMapMemoryInBytes()
         );
-
     }
 
     private ClusterStatsNodeResponse createClusterStatsNodeResponse(